requirements.txt
# Top-level dependencies only - pip will resolve sub-dependencies
pandas==2.2.3
requests==2.32.3
biopython==1.85
python-dotenv==1.0.0
flask==3.0.0
gunicorn==21.2.0
